delete and move some items
